Add <Suspense> boundary and localStorage fallback
Using specific dynamic components in Next.js (in this case,
`useSearchParams()`) causes the entire tree above to be rendered
on the client, effectively disabling SSR for the entire site. To
restore SSR, we add a <Suspense> boundary around the offending
component (SearchAndFilter) and ensure all client components
continue to work without access to localStorage.